Added more useful output when the serialiser can't handle an ivar.
Added category that serialises / deserialises GSMutableString except for the NSZone.  NSZones are probably going to want some special (but generic) treatment.

WARNING: The version of GSPrivate.h, the GSMutableString+ETSerialisable category, and the installed version of GNUstep must /all/ be kept in sync.  In future, some preprocessor magic will be used to do this automatically, but it hasn't been done yet.