Deleted accidentally committed file (what are all these .d files anyway?).
Added versioning support.  Objects automatically have their class version stored with them.  The deserialise:withPointer: method now has a version: parameter.  This can be used to implement sensible behaviour when changing the layout of a class.  Remember: Call +setVersion: on your class if you change the ivars...

Category on GSMutableString now completed more-or-less, although it does lose the zone information.  This is almost certainly a good thing, since zones contain pointers to functions that might not exist when the object is reloaded.  At some point, I should implement some mechanism for deserialising a graph into a specific zone.  This could be very useful for creating temporary zones that can be freed all in one go to bin an object graph that was deserialised temporarily.