A few EtoileSerialise bug fixes.  Framework now works well enough to serialise an NSTextStorage object and messages sent to it.  I have a version of TypeWriter that inserts a COProxy in between the NSTextStorage object and the rest of the text system, giving a complete history of every single message that has been sent into the object.
COProxy now automatically uses the correct subclass.  If you instantiate a COProxy for an instance of Foo then it will attempt to use the COProxy_Foo class.  If this doesn't exist it will keep looking up the class hierarchy until it finds one that works.  A COProxy_NSAttributedString has been added, which doesn't log any of the immutable methods for NSAttributedString.  Copy this for your own COProxy subclasses.