Lots of changes to EtoileSerialise.  It turns out the GS[Mutable]String code was a good approximation of nonsense and only passed the tests through pure chance.  It is now a much worse approximation of nonsense.
The big news, however, is that COProxy now fully supports playback.  You can create an object, wrap it in a COProxy, send it messages, and then (at any later point) send it a -setVersion: message and get an earlier version of the object back.  Lots of bits in the rest of the framework have been poked to make this work.