Merge CoreObject r3178:3820 from trunk/Etoile/Frameworks/CoreObject into stable/Frameworks/CoreObject
This includes all the work done on CoreObject over the past twelve days.
CoreObject in stable is still work-in-progress; undo/redo and COProxy are part 
of this merge but still untested. I'm pushing the merge now to get testing for 
the other changes and also to fix CoreObject compilation in stable on platforms 
where custom library and include flags are needed for libpq (see 
http://review.etoileos.com/r/36/).

WARNING: The DB schema has changed, you must run uninstall-coreobject.sh before 
being able to use CoreObject if it was installed previously. All core objects 
trace will be lost in the DB.