Added support for static compilation to LanguageKit.  Modules can now be compiled to LLVM bitcode which is written to a file.
Modified edlc to support -c for static compilation.  -c was previously used for specifying the class to instantiate.  -C is now used for this (-c is to be consistent with other compilers). 

Added native example, where test.st is statically compiled and linked into a binary.  This is currently an ugly hack, but will be nicer once GNUstep Make supports Smalltalk (which it couldn't previously because the compiler didn't).