New event handling model - part 1
Rewritten ETEvent to be a standalone event class that wraps backend events 
rather than an NSEvent subclass. This simplifies the code a bit, makes possible 
to cleanly synthetize our own events in the other code parts to be committed and 
paves the way to support other widget backends than AppKit in future (but we are 
far from there... :-).