Fixed compilation when Foundation.h is included.
Added implementation of blocks.  Much of this code is by Remy Demarest.  The original can be found here: http://github.com/PsychoH13/C-ObjC-Blocks/tree/master.  It is incorporated here with the knowledge and consent of the original author, and was originally released under the same MIT license as this framework.