Fixed hit test broken in every item which corresponds to a window content view.
The issue was only visible when a custom layout was set on a layout item 
decorated with a window.

The hit item was correctly returned but the hit test location in the event was 
getting wrongly updated with -[[hitItem layout] rootItem]. -hitTestWithEvent: 
was not verifying as expected whether the hit item accepts actions or not, and 
just updating the event in all cases.