Added new ETStyleGroup class. First step in making the styling support really usable and clean.
The current version implements styles as linked list which can bound to a layout 
item with -setStyle:. However this doesn't work well because styles tend to be 
shared and you get in trouble when you want to chain new styles without altering 
all the items which share the existing style. See ChangeLog also.