Renamed -render:dirtyRect:inView: to -render:dirtyRect:inContext:. Renamed private method display:item:dirtyRect:inView: to -display:item:dirtyRect:inContext:.
Now we can pass an arbitrary rendering context like Web, GL or pseudo 3D to 
apply additional transforms or alter how NSGraphicsContext operations are 
handled/interpreted. Having a view as parameter proved to be completely 
useless, if we want to retarget our drawing to a specific view, it's enough to 
just lock the focus in this view before invoking the rendering of the layout 
item tree.