Fixed the ubuntu install script and improved it to support options such as prefix, build directory and built version (trunk or stable).
Added a help section that can be accessed with --help.

setup.sh is now run automatically at the end, which means the resulting 
installation is fully functional and the environment can be started by picking 
the Etoile session in GDM.

Script successfully tested with trunk r4894. Building stable is untested, but 
should work well, reports are welcome :-)