Rewritten -handleRemove:items: and -handleAdd:items to use FOREACH macro.
Fixed -handleRemove:item: to check that the item to remove really belongs to the 
receiver before removing it. Until now having no such check means it was easy 
to remove an item belonging to another parent and corrupts it in a way which 
later leads to a weird crash (_parentItem being nil when it shouldn't).

Improved -handleInsert:item:atIndex: and -handleAdd:item: to check whether 
the item already belongs to the receiver in a less slow way. We now  test 
whether the parent is equal to the receiver (exactly as we do in 
-handleRemove:item:) rather than whether the item is a receiver child.