Made adding and removing large collection of items a lot faster when autolayout is enabled.
Previously with autolayout enabled, each insertion or removal would trigger a 
layout update, even when it was invoked by -addItems: or -removeItems:.

Added -beginMutate and -endMutate: to suppor that in ETLayoutItemGroup+Mutation.

Added -tryReloadWithSource: to consolidate the reloading logic in a way that 
allows -setSource: to use it.
Fixed -setSource: to call -tryReloadWithSource: in order to reload and 
update the layout as the documentation states, and finally eliminates 
-removeAllItems: which was triggering an unecessary layout update.