Renamed -setLayoutItem: to -setItem: in ETView to take in account the parameter can be an ETUIItem now and not only an ETLayoutItem.
Renamed _layoutItem ivar to 'item' and turned it into an IBOutlet to support 
better integration with Gorm/IB. Will be used as a replacement to instantiate 
an ETContainer as custom view.

Updated code in other classes that were using -setLayoutItem:. 
Updated ETView/ETContainer code which were using _layoutItem or 
variables/arguments named 'item'.