Overrided -saveAndOverrideAutoresizingMaskOfDecoratedItem: and -restoreAutoresizingMaskOfDecoratedItem: in ETWindowItem to support restoring the decorated item exactly as it was.
Added -convertRectToWidgetBackendScreenBase: and 
-convertRectFromWidgetBackendScreenBase: by extracting -decorationRect code.

Fixed -handleSetDecorationRect: to use the rect parameter rather than ignoring 
it and just reuse the existing window origin.
Fixed -handleDecorateItem:supervisorView:inView: similarly.

Tweaked -renderWindow: in ETEtoileUIBuilder to keep the existing window frame 
when an AppKit window is converted to a window item (starting from now this is 
not the default ETWindowItem behavior anymore).