Largely rewrote ETIconLayout to leverage existing pluggable aspects (ETBasicItemStyle, ETActionHandler etc.) by subclassing them. The icon & label look is much more polished and label editing is now usable.
Modified ETEventProcess to remove the field editor on a first responder change, 
in case the field editor is managed by EtoileUI rather than the AppKit.

Improved -mouseDown: in instrument classes to remove the field editor item when 
the event hasn't been delegated to it.

Fixed ETArrowTool and ETMoveTool to mark the event as delivered when it just 
got turned into an action.

Fixed ETSelectTool to dispatch key down events normally when the first responder 
is not the tool itself.