More tidying.
Note: Testing this code showed some bugs in clang (my fault - sorry!), which are fixed in r101759.  It is strongly recommended that you upgrade clang to this revision or newer.  There is also a bug in mainline clang preventing his from compiling.  You can work around it, but it's worth waiting for Douglas to fix it properly.