Some interpreter bug fixes:
- Correctly box smallint returns (e.g. from BigInt).  
- Correctly handle variable references from blocks.

With these fixes, the LK interpreter can now run test.st without crashing.  It takes about 2/3 as much CPU time as the compiler, although about 70% of the total run time is doing I/O, so that's not a huge gain, but it shows that the interpreter is a good choice for short-lived scripts and programs.  The interpreter uses a peak of 20MB of RAM, while the compiler uses 35MB, so there's also a saving there, although there are probably some leaks in the interpreter...